requirements.txt
python-dotenv==1.0.1
langchain==0.3.21
langchain-core==0.3.45
langchain-openai==0.3.9
langchain-text-splitters==0.3.7
langsmith==0.3.16
langchain-pinecone==0.2.3
langchain-community==0.3.20
openai==1.66.3
langchain_chroma==0.2.2
langchain-google-genai==2.0.6